requirements.txt
absl-py==2.3.1
accelerate==1.7.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiosignal==1.4.0
altair==5.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
cachetools==6.2.1
certifi==2024.7.4
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
deepspeed==0.17.1
diffusers==0.35.1
einops==0.6.1
exceptiongroup==1.3.0
fastapi==0.120.0
ffmpy==0.6.4
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
glcontext==3.0.0
google-auth==2.41.1
google-auth-oauthlib==1.0.0
gradio==5.17.1
gradio_client==1.7.1
grpcio==1.66.0
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==0.36.0
idna==3.11
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib_metadata==8.7.0
inquirerpy==0.3.4
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
lightning-utilities==0.15.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moderngl==5.12.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
narwhals==2.9.0
networkx==3.4.2
ninja==1.13.0
numpy==1.23.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
omegaconf==2.3.0
open-clip-torch==2.17.1
opencv-python==4.7.0.72
OpenEXR==3.4.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
peft==0.17.0
pfzy==0.3.4
Pillow==9.4.0
platformdirs==4.5.0
plyfile==1.1.3
prodigyopt==1.1.2
proglog==0.1.12
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.1.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
pyglet==1.5.0
Pygments==2.18.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==1.9.3
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
roma==1.5.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.2
safehttpx==0.1.7
safetensors==0.6.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.42.1
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.40.0
sympy==1.14.0
tensorboard==2.12.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.6.0
tifffile==2024.8.24
timm==1.0.15
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.7.0
torchmetrics==1.8.2
torchvision==0.22.0
tqdm==4.65.0
transformers==4.52.3
trimesh==4.4.3
triton==3.3.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@c5daf6f6c244d251f252102d09e9b7bcef791a38
uvicorn==0.38.0
wandb==0.20.1
wcwidth==0.2.14
websockets==14.2
Werkzeug==3.1.3
xformers==0.0.30
yarl==1.9.4
zipp==3.23.0
